[feature] Added forgotten file in the previous large commit with changes: “src/agents/components”.
In the file, the sizes and shapes of the different tensors are now handled correctly. In addition, a modified version of the “forward” method of the ActorCriticPolicy class was added to the custom policy.

The commit is leveraged to add information on how to run a test training in the README. In addition, the file names in the “./experiments/scripts” were changed.